per Damian-Nordic, revert networkIDLen to uint8_t
But, use static_cast<uint8_t> to fix build warning

+++ src/platform/Linux/NetworkCommissioningDriver.h
@@ -180,8 +180,8 @@ public:
 
         // Public, but cannot be accessed via NetworkIterator interface.
         uint8_t interfaceName[kMaxNetworkIDLen];
-        size_t interfaceNameLen = 0;
-        bool exhausted          = false;
+        uint8_t interfaceNameLen = 0;
+        bool exhausted           = false;
     };
 
     uint8_t GetMaxNetworks() override { return 1; };
